+++ Dungeon.csv
@@ -13,7 +13,7 @@
 -1,-1,-1,0,-1,-1,-1,0,-1,-1,0,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1
 -1,0,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1,-1
 -1,0,-1,0,-1,-1,0,0,0,-1,-1,-1,-1,-1,-1,-1,-1,0,-1,-1,0,-1,-1,-1
-0,0,0,0,-1,-1,-1,0,-1,-1,-1,-1,-1,-1,-1,-1,0,0,0,0,0,0,0,-1
+0,0,0,0,-1,c,-1,0,-1,-1,-1,-1,-1,-1,-1,c,0,0,0,0,0,0,0,-1
 -1,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,0,0,-1,-1,0,-1,-1,-1,-1,-1,0,-1
 -1,-1,-1,0,-1,-1,-1,0,-1,c,-1,-1,0,-1,-1,-1,0,-1,-1,-1,-1,-1,0,-1
 -1,-1,-1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,-1,-1,-1
