+++ imborjon-Dungeon.csv
@@ -1,0 +1,20 @@
+﻿-1,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,|,|,|,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,|,|,|,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,|,|,|,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,0,0,|,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,|,|,|,0,|,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,0,0,|,|,0,0,0,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,0,|,|,0,|,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,0,|,|,0,|,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,|,|,|,0,0,0,0,0,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,0,|,|,|,0,|,|,0,|,0,|,|,|,|,|,|,|,|,|,|,|
+|,|,|,0,|,|,|,0,|,|,0,0,0,0,|,|,|,|,|,|,|,|,|,|
+|,0,|,0,|,|,|,0,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|
+|,0,|,0,|,|,0,0,0,|,|,|,|,|,|,|,|,0,|,|,0,|,|,|
+0,0,0,0,|,|,|,0,|,|,|,|,|,|,|,|,0,0,0,0,0,0,0,|
+|,|,|,$,|,|,|,0,|,|,|,|,0,0,|,|,0,|,|,|,|,|,0,|
+|,|,|,0,|,|,|,0,|,c,|,|,0,|,|,|,0,|,|,|,|,|,0,|
+|,|,|,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,|,|,|
+|,|,|,|,|,|,|,|,e,|,|,|,|,|,|,|,|,|,|,|,|,|,|,|
