+++ imborjonPlayer.csv
@@ -1,0 +1,14 @@
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
+HealthPoints,damage,heal_potion
+100,7.0,3
