+++ log.txt
@@ -1,0 +1,101 @@
+Player: null logged out Sun Nov 12 23:18:49 MST 2023 
+Player: null logged in  Sun Nov 12 23:19:00 MST 2023 
+Player: null logged out Sun Nov 12 23:19:12 MST 2023 
+Player: null logged in  Sun Nov 12 23:19:16 MST 2023 
+Player: null logged out Sun Nov 12 23:19:33 MST 2023 
+Player: null logged in  Sun Nov 12 23:19:36 MST 2023 
+Player: null logged out Sun Nov 12 23:19:45 MST 2023 
+Player: null logged in  Sun Nov 12 23:19:46 MST 2023 
+Player: null logged out Sun Nov 12 23:19:58 MST 2023 
+Player: null logged in  Sun Nov 12 23:20:00 MST 2023 
+Player: null logged out Sun Nov 12 23:20:12 MST 2023 
+Player: null logged in  Sun Nov 12 23:20:13 MST 2023 
+Player: null logged out Sun Nov 12 23:20:54 MST 2023 
+Player: null logged in  Sun Nov 12 23:21:10 MST 2023 
+Player: null logged out Sun Nov 12 23:22:59 MST 2023 
+Player: null logged out Sun Nov 12 23:23:06 MST 2023 
+Player: null logged out Sun Nov 12 23:24:10 MST 2023 
+Player: imborjon logged in  Sun Nov 12 23:24:38 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Sun Nov 12 23:25:07 MST 2023 
+imborjon Moved Left, This is the current position: 18,7 Sun Nov 12 23:25:10 MST 2023 
+imborjon Moved Left, This is the current position: 18,6 Sun Nov 12 23:25:10 MST 2023 
+Player: imborjon logged out Sun Nov 12 23:25:13 MST 2023 
+Player: imborjon logged in  Sun Nov 12 23:25:20 MST 2023 
+imborjon Moved UP, This is the current position: 18,6 Sun Nov 12 23:25:23 MST 2023 
+VenomPlant killed imborjon in battle Sun Nov 12 23:26:45 MST 2023 
+Player: izuckDick Register in to the game  Sun Nov 12 23:28:48 MST 2023 
+Player: izuckDick logged in  Sun Nov 12 23:29:13 MST 2023 
+Player: izuckDick started a new game  Sun Nov 12 23:29:13 MST 2023 
+izuckDick Moved UP, This is the current position: 18,8 Sun Nov 12 23:29:49 MST 2023 
+izuckDick Moved Left, This is the current position: 18,7 Sun Nov 12 23:29:51 MST 2023 
+izuckDick Moved Left, This is the current position: 18,6 Sun Nov 12 23:29:51 MST 2023 
+izuckDick Moved Left, This is the current position: 18,5 Sun Nov 12 23:29:52 MST 2023 
+izuckDick Moved Left, This is the current position: 18,4 Sun Nov 12 23:29:54 MST 2023 
+izuckDick Moved Left, This is the current position: 18,3 Sun Nov 12 23:29:55 MST 2023 
+Snake killed izuckDick in battle Sun Nov 12 23:30:06 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:04:39 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:04:43 MST 2023 
+imborjon Moved Left, This is the current position: 18,7 Mon Nov 13 00:04:43 MST 2023 
+imborjon Moved Left, This is the current position: 18,6 Mon Nov 13 00:04:44 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:08:02 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:08:04 MST 2023 
+imborjon Moved Right, This is the current position: 18,9 Mon Nov 13 00:08:14 MST 2023 
+imborjon Moved Right, This is the current position: 18,10 Mon Nov 13 00:08:17 MST 2023 
+imborjon Moved Right, This is the current position: 18,11 Mon Nov 13 00:08:23 MST 2023 
+imborjon Moved Right, This is the current position: 18,12 Mon Nov 13 00:08:25 MST 2023 
+imborjon Moved Right, This is the current position: 18,13 Mon Nov 13 00:08:27 MST 2023 
+imborjon Moved Right, This is the current position: 18,14 Mon Nov 13 00:08:27 MST 2023 
+imborjon Moved Right, This is the current position: 18,15 Mon Nov 13 00:08:28 MST 2023 
+imborjon Moved Left, This is the current position: 18,14 Mon Nov 13 00:08:32 MST 2023 
+imborjon Moved Left, This is the current position: 18,13 Mon Nov 13 00:08:32 MST 2023 
+imborjon Moved Left, This is the current position: 18,12 Mon Nov 13 00:08:33 MST 2023 
+imborjon Moved Left, This is the current position: 18,11 Mon Nov 13 00:08:33 MST 2023 
+imborjon Moved Left, This is the current position: 18,10 Mon Nov 13 00:08:34 MST 2023 
+imborjon Moved Left, This is the current position: 18,9 Mon Nov 13 00:08:38 MST 2023 
+imborjon Moved Left, This is the current position: 18,8 Mon Nov 13 00:08:39 MST 2023 
+imborjon Moved Left, This is the current position: 18,7 Mon Nov 13 00:08:46 MST 2023 
+imborjon Moved Left, This is the current position: 18,6 Mon Nov 13 00:08:47 MST 2023 
+imborjon Moved Left, This is the current position: 18,5 Mon Nov 13 00:08:47 MST 2023 
+imborjon Moved Left, This is the current position: 18,4 Mon Nov 13 00:08:48 MST 2023 
+imborjon Moved Left, This is the current position: 18,3 Mon Nov 13 00:08:48 MST 2023 
+Goblin killed imborjon in battle Mon Nov 13 00:08:50 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:12:06 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:12:08 MST 2023 
+imborjon Moved Left, This is the current position: 18,7 Mon Nov 13 00:12:13 MST 2023 
+imborjon Moved Left, This is the current position: 18,6 Mon Nov 13 00:12:14 MST 2023 
+imborjon Moved Left, This is the current position: 18,5 Mon Nov 13 00:12:18 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:14:59 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:15:07 MST 2023 
+imborjon Moved Right, This is the current position: 18,9 Mon Nov 13 00:15:13 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:16 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:17 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:18 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:19 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:19 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:20 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:21 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:21 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:22 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:22 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:23 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:24 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:25 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:26 MST 2023 
+imborjon Moved UP, This is the current position: 18,9 Mon Nov 13 00:15:27 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:27:43 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:27:46 MST 2023 
+imborjon Moved Right, This is the current position: 18,9 Mon Nov 13 00:27:47 MST 2023 
+Player: imborjon logged in  Mon Nov 13 00:28:15 MST 2023 
+imborjon Moved UP, This is the current position: 18,8 Mon Nov 13 00:28:17 MST 2023 
+imborjon Moved Right, This is the current position: 18,9 Mon Nov 13 00:28:17 MST 2023 
+imborjon Moved UP, This is the current position: 17,9 Mon Nov 13 00:28:19 MST 2023 
+imborjon Moved Down, This is the current position: 18,9 Mon Nov 13 00:28:33 MST 2023 
+imborjon Moved Left, This is the current position: 18,8 Mon Nov 13 00:28:34 MST 2023 
+imborjon Moved Left, This is the current position: 18,7 Mon Nov 13 00:28:55 MST 2023 
+imborjon Moved Left, This is the current position: 18,6 Mon Nov 13 00:28:58 MST 2023 
+imborjon Moved Left, This is the current position: 18,5 Mon Nov 13 00:28:58 MST 2023 
+imborjon Moved Left, This is the current position: 18,4 Mon Nov 13 00:29:03 MST 2023 
+imborjon Moved Left, This is the current position: 18,3 Mon Nov 13 00:29:04 MST 2023 
+imborjon Moved UP, This is the current position: 17,3 Mon Nov 13 00:29:05 MST 2023 
+imborjon Moved UP, This is the current position: 16,3 Mon Nov 13 00:29:05 MST 2023 
+DarkKing killed imborjon in battle Mon Nov 13 00:29:09 MST 2023
